requirements/base.txt
Django              ==2.0.2
# django-mobile official repository only supports up to Django 1.9
git+git://github.com/frgmt/django-mobile#egg=django-mobile

psycopg2-binary     ==2.7.5
pyproj              ==1.9.5.1
django-appconf      ==1.0.2
py-bcrypt           ==0.4
Pillow              ==4.2.1
httplib2            ==0.9.1
beautifulsoup4      ==4.6.0
cchardet            ==2.1.1
# linaro-django-pagination official repository does not support Django 2.0.
git+git://github.com/frgmt/django-pagination#egg=linaro-django-pagination
django-encrypted-fields-python3 ==1.1.3
user-agents         ==1.1.0
pyyaml              ==3.12
requests            ==2.19.1
six                 ==1.11.0
slacker             ==0.9.65

# time
pytz                ==2018.5
python-dateutil     ==2.6.1

# aws
boto                ==2.48.0
boto3               ==1.7.50

# upload static files to S3
django-storages     ==1.6.5

# log
logutils            ==0.3.3
raven               ==6.9.0

# api
djangorestframework ==3.7.7
django-cors-headers ==2.2.0
django-filter       ==1.0.4
django-admin-rangefilter ==0.3.9
# official django-subdomains seems to have been abandoned by the maintainer
git+git://github.com/frgmt/django-subdomains#egg=django-subdomains
django-rest-swagger ==2.1.2
django-bulk-update  ==2.2.0
rauth               ==0.7.3

# twitter api
twitter             ==1.17.1

# redis
redis == 2.10.6
hiredis == 0.2.0
django-redis == 4.9.0
